# artifacts/api-server/src/routes/bookings.ts
import { Router } from "express";
import { db, bookingsTable } from "@workspace/db";
import { eq, and } from "drizzle-orm";
import { ListBookingsQueryParams, CreateBookingBody, UpdateBookingParams, UpdateBookingBody } from "@workspace/api-zod";

const router = Router();

router.get("/bookings", async (req, res) => {
  const query = ListBookingsQueryParams.safeParse(req.query);
  if (!query.success) {
    return res.status(400).json({ error: "Invalid query params" });
  }
  const { artisanId, status } = query.data;

  const conditions = [];
  if (artisanId) conditions.push(eq(bookingsTable.artisanId, artisanId));
  if (status) conditions.push(eq(bookingsTable.status, status as "pending" | "confirmed" | "completed" | "cancelled"));

  const where = conditions.length > 0 ? and(...conditions) : undefined;
  const bookings = await db.select().from(bookingsTable).where(where);

  res.json({ bookings, total: bookings.length });
});

router.post("/bookings", async (req, res) => {
  const body = CreateBookingBody.safeParse(req.body);
  if (!body.success) {
    return res.status(400).json({ error: "Invalid body" });
  }
  const [booking] = await db.insert(bookingsTable).values(body.data).returning();
  res.status(201).json(booking);
});

router.put("/bookings/:id", async (req, res) => {
  const params = UpdateBookingParams.safeParse({ id: Number(req.params.id) });
  const body = UpdateBookingBody.safeParse(req.body);
  if (!params.success || !body.success) {
    return res.status(400).json({ error: "Invalid request" });
  }
  const [booking] = await db
    .update(bookingsTable)
    .set({ status: body.data.status as "pending" | "confirmed" | "completed" | "cancelled" })
    .where(eq(bookingsTable.id, params.data.id))
    .returning();
  if (!booking) return res.status(404).json({ error: "Booking not found" });
  res.json(booking);
});

export default router;