# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ArtisanSN - Marketplace de Services API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Artisan {
  id: number;
  name: string;
  bio?: string;
  category: string;
  city: string;
  phone: string;
  email?: string;
  photoUrl?: string;
  priceRange?: string;
  yearsExperience?: number;
  isVerified: boolean;
  averageRating: number;
  reviewCount: number;
  createdAt: string;
}

export interface ArtisanList {
  artisans: Artisan[];
  total: number;
}

export interface CreateArtisanBody {
  name: string;
  bio?: string;
  category: string;
  city: string;
  phone: string;
  email?: string;
  photoUrl?: string;
  priceRange?: string;
  yearsExperience?: number;
}

export interface UpdateArtisanBody {
  name?: string;
  bio?: string;
  category?: string;
  city?: string;
  phone?: string;
  email?: string;
  photoUrl?: string;
  priceRange?: string;
  yearsExperience?: number;
}

export interface ServiceCategory {
  id: number;
  name: string;
  icon: string;
  description?: string;
  artisanCount: number;
}

export interface ServiceList {
  services: ServiceCategory[];
}

export interface Review {
  id: number;
  artisanId: number;
  clientName: string;
  rating: number;
  comment?: string;
  createdAt: string;
}

export interface ReviewList {
  reviews: Review[];
  total: number;
}

export interface CreateReviewBody {
  artisanId: number;
  clientName: string;
  rating: number;
  comment?: string;
}

export type BookingStatus = (typeof BookingStatus)[keyof typeof BookingStatus];

export const BookingStatus = {
  pending: "pending",
  confirmed: "confirmed",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface Booking {
  id: number;
  artisanId: number;
  clientName: string;
  clientPhone: string;
  serviceDescription: string;
  scheduledDate?: string;
  status: BookingStatus;
  city?: string;
  createdAt: string;
}

export interface BookingList {
  bookings: Booking[];
  total: number;
}

export interface CreateBookingBody {
  artisanId: number;
  clientName: string;
  clientPhone: string;
  serviceDescription: string;
  scheduledDate?: string;
  city?: string;
}

export type UpdateBookingBodyStatus =
  (typeof UpdateBookingBodyStatus)[keyof typeof UpdateBookingBodyStatus];

export const UpdateBookingBodyStatus = {
  pending: "pending",
  confirmed: "confirmed",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface UpdateBookingBody {
  status: UpdateBookingBodyStatus;
}

export interface PlatformStats {
  totalArtisans: number;
  totalBookings: number;
  totalCities: number;
  totalCategories: number;
  completedJobs: number;
  averageRating: number;
}

export type ListArtisansParams = {
  category?: string;
  city?: string;
  search?: string;
  limit?: number;
  offset?: number;
};

export type ListReviewsParams = {
  artisanId: number;
};

export type ListBookingsParams = {
  artisanId?: number;
  status?: ListBookingsStatus;
};

export type ListBookingsStatus =
  (typeof ListBookingsStatus)[keyof typeof ListBookingsStatus];

export const ListBookingsStatus = {
  pending: "pending",
  confirmed: "confirmed",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export type GetTopRatedArtisansParams = {
  limit?: number;
};