# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ArtisanSN - Marketplace de Services API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: artisans
    description: Artisan profiles and management
  - name: services
    description: Service categories
  - name: reviews
    description: Client reviews
  - name: bookings
    description: Service bookings

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /artisans:
    get:
      operationId: listArtisans
      tags: [artisans]
      summary: List artisans with optional filters
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: city
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of artisans
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtisanList"
    post:
      operationId: createArtisan
      tags: [artisans]
      summary: Register a new artisan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateArtisanBody"
      responses:
        "201":
          description: Artisan created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artisan"

  /artisans/{id}:
    get:
      operationId: getArtisan
      tags: [artisans]
      summary: Get artisan by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Artisan details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artisan"
        "404":
          description: Not found
    put:
      operationId: updateArtisan
      tags: [artisans]
      summary: Update artisan profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateArtisanBody"
      responses:
        "200":
          description: Updated artisan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artisan"

  /services:
    get:
      operationId: listServices
      tags: [services]
      summary: List all service categories
      responses:
        "200":
          description: Service categories
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceList"

  /reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      summary: List reviews for an artisan
      parameters:
        - name: artisanId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of reviews
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewList"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Create a review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewBody"
      responses:
        "201":
          description: Review created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List bookings
      parameters:
        - name: artisanId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, completed, cancelled]
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingList"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingBody"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"

  /bookings/{id}:
    put:
      operationId: updateBooking
      tags: [bookings]
      summary: Update booking status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBookingBody"
      responses:
        "200":
          description: Updated booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"

  /stats/summary:
    get:
      operationId: getStats
      tags: [artisans]
      summary: Platform summary stats
      responses:
        "200":
          description: Summary stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformStats"

  /artisans/top-rated:
    get:
      operationId: getTopRatedArtisans
      tags: [artisans]
      summary: Get top rated artisans
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 6
      responses:
        "200":
          description: Top rated artisans
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtisanList"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Artisan:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        bio:
          type: string
        category:
          type: string
        city:
          type: string
        phone:
          type: string
        email:
          type: string
        photoUrl:
          type: string
        priceRange:
          type: string
        yearsExperience:
          type: integer
        isVerified:
          type: boolean
        averageRating:
          type: number
        reviewCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - category
        - city
        - phone
        - isVerified
        - averageRating
        - reviewCount
        - createdAt

    ArtisanList:
      type: object
      properties:
        artisans:
          type: array
          items:
            $ref: "#/components/schemas/Artisan"
        total:
          type: integer
      required:
        - artisans
        - total

    CreateArtisanBody:
      type: object
      properties:
        name:
          type: string
        bio:
          type: string
        category:
          type: string
        city:
          type: string
        phone:
          type: string
        email:
          type: string
        photoUrl:
          type: string
        priceRange:
          type: string
        yearsExperience:
          type: integer
      required:
        - name
        - category
        - city
        - phone

    UpdateArtisanBody:
      type: object
      properties:
        name:
          type: string
        bio:
          type: string
        category:
          type: string
        city:
          type: string
        phone:
          type: string
        email:
          type: string
        photoUrl:
          type: string
        priceRange:
          type: string
        yearsExperience:
          type: integer

    ServiceCategory:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        icon:
          type: string
        description:
          type: string
        artisanCount:
          type: integer
      required:
        - id
        - name
        - icon
        - artisanCount

    ServiceList:
      type: object
      properties:
        services:
          type: array
          items:
            $ref: "#/components/schemas/ServiceCategory"
      required:
        - services

    Review:
      type: object
      properties:
        id:
          type: integer
        artisanId:
          type: integer
        clientName:
          type: string
        rating:
          type: integer
        comment:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - artisanId
        - clientName
        - rating
        - createdAt

    ReviewList:
      type: object
      properties:
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/Review"
        total:
          type: integer
      required:
        - reviews
        - total

    CreateReviewBody:
      type: object
      properties:
        artisanId:
          type: integer
        clientName:
          type: string
        rating:
          type: integer
        comment:
          type: string
      required:
        - artisanId
        - clientName
        - rating

    Booking:
      type: object
      properties:
        id:
          type: integer
        artisanId:
          type: integer
        clientName:
          type: string
        clientPhone:
          type: string
        serviceDescription:
          type: string
        scheduledDate:
          type: string
          format: date-time
        status:
          type: string
          enum: [pending, confirmed, completed, cancelled]
        city:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - artisanId
        - clientName
        - clientPhone
        - serviceDescription
        - status
        - createdAt

    BookingList:
      type: object
      properties:
        bookings:
          type: array
          items:
            $ref: "#/components/schemas/Booking"
        total:
          type: integer
      required:
        - bookings
        - total

    CreateBookingBody:
      type: object
      properties:
        artisanId:
          type: integer
        clientName:
          type: string
        clientPhone:
          type: string
        serviceDescription:
          type: string
        scheduledDate:
          type: string
          format: date-time
        city:
          type: string
      required:
        - artisanId
        - clientName
        - clientPhone
        - serviceDescription

    UpdateBookingBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, completed, cancelled]
      required:
        - status

    PlatformStats:
      type: object
      properties:
        totalArtisans:
          type: integer
        totalBookings:
          type: integer
        totalCities:
          type: integer
        totalCategories:
          type: integer
        completedJobs:
          type: integer
        averageRating:
          type: number
      required:
        - totalArtisans
        - totalBookings
        - totalCities
        - totalCategories
        - completedJobs
        - averageRating