# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ArtisanSN - Marketplace de Services API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List artisans with optional filters
 */
export const listArtisansQueryLimitDefault = 20;
export const listArtisansQueryOffsetDefault = 0;

export const ListArtisansQueryParams = zod.object({
  category: zod.coerce.string().optional(),
  city: zod.coerce.string().optional(),
  search: zod.coerce.string().optional(),
  limit: zod.coerce.number().default(listArtisansQueryLimitDefault),
  offset: zod.coerce.number().default(listArtisansQueryOffsetDefault),
});

export const ListArtisansResponse = zod.object({
  artisans: zod.array(
    zod.object({
      id: zod.number(),
      name: zod.string(),
      bio: zod.string().optional(),
      category: zod.string(),
      city: zod.string(),
      phone: zod.string(),
      email: zod.string().optional(),
      photoUrl: zod.string().optional(),
      priceRange: zod.string().optional(),
      yearsExperience: zod.number().optional(),
      isVerified: zod.boolean(),
      averageRating: zod.number(),
      reviewCount: zod.number(),
      createdAt: zod.coerce.date(),
    }),
  ),
  total: zod.number(),
});

/**
 * @summary Register a new artisan
 */
export const CreateArtisanBody = zod.object({
  name: zod.string(),
  bio: zod.string().optional(),
  category: zod.string(),
  city: zod.string(),
  phone: zod.string(),
  email: zod.string().optional(),
  photoUrl: zod.string().optional(),
  priceRange: zod.string().optional(),
  yearsExperience: zod.number().optional(),
});

/**
 * @summary Get artisan by ID
 */
export const GetArtisanParams = zod.object({
  id: zod.coerce.number(),
});

export const GetArtisanResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  bio: zod.string().optional(),
  category: zod.string(),
  city: zod.string(),
  phone: zod.string(),
  email: zod.string().optional(),
  photoUrl: zod.string().optional(),
  priceRange: zod.string().optional(),
  yearsExperience: zod.number().optional(),
  isVerified: zod.boolean(),
  averageRating: zod.number(),
  reviewCount: zod.number(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Update artisan profile
 */
export const UpdateArtisanParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateArtisanBody = zod.object({
  name: zod.string().optional(),
  bio: zod.string().optional(),
  category: zod.string().optional(),
  city: zod.string().optional(),
  phone: zod.string().optional(),
  email: zod.string().optional(),
  photoUrl: zod.string().optional(),
  priceRange: zod.string().optional(),
  yearsExperience: zod.number().optional(),
});

export const UpdateArtisanResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  bio: zod.string().optional(),
  category: zod.string(),
  city: zod.string(),
  phone: zod.string(),
  email: zod.string().optional(),
  photoUrl: zod.string().optional(),
  priceRange: zod.string().optional(),
  yearsExperience: zod.number().optional(),
  isVerified: zod.boolean(),
  averageRating: zod.number(),
  reviewCount: zod.number(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary List all service categories
 */
export const ListServicesResponse = zod.object({
  services: zod.array(
    zod.object({
      id: zod.number(),
      name: zod.string(),
      icon: zod.string(),
      description: zod.string().optional(),
      artisanCount: zod.number(),
    }),
  ),
});

/**
 * @summary List reviews for an artisan
 */
export const ListReviewsQueryParams = zod.object({
  artisanId: zod.coerce.number(),
});

export const ListReviewsResponse = zod.object({
  reviews: zod.array(
    zod.object({
      id: zod.number(),
      artisanId: zod.number(),
      clientName: zod.string(),
      rating: zod.number(),
      comment: zod.string().optional(),
      createdAt: zod.coerce.date(),
    }),
  ),
  total: zod.number(),
});

/**
 * @summary Create a review
 */
export const CreateReviewBody = zod.object({
  artisanId: zod.number(),
  clientName: zod.string(),
  rating: zod.number(),
  comment: zod.string().optional(),
});

/**
 * @summary List bookings
 */
export const ListBookingsQueryParams = zod.object({
  artisanId: zod.coerce.number().optional(),
  status: zod
    .enum(["pending", "confirmed", "completed", "cancelled"])
    .optional(),
});

export const ListBookingsResponse = zod.object({
  bookings: zod.array(
    zod.object({
      id: zod.number(),
      artisanId: zod.number(),
      clientName: zod.string(),
      clientPhone: zod.string(),
      serviceDescription: zod.string(),
      scheduledDate: zod.coerce.date().optional(),
      status: zod.enum(["pending", "confirmed", "completed", "cancelled"]),
      city: zod.string().optional(),
      createdAt: zod.coerce.date(),
    }),
  ),
  total: zod.number(),
});

/**
 * @summary Create a booking
 */
export const CreateBookingBody = zod.object({
  artisanId: zod.number(),
  clientName: zod.string(),
  clientPhone: zod.string(),
  serviceDescription: zod.string(),
  scheduledDate: zod.coerce.date().optional(),
  city: zod.string().optional(),
});

/**
 * @summary Update booking status
 */
export const UpdateBookingParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateBookingBody = zod.object({
  status: zod.enum(["pending", "confirmed", "completed", "cancelled"]),
});

export const UpdateBookingResponse = zod.object({
  id: zod.number(),
  artisanId: zod.number(),
  clientName: zod.string(),
  clientPhone: zod.string(),
  serviceDescription: zod.string(),
  scheduledDate: zod.coerce.date().optional(),
  status: zod.enum(["pending", "confirmed", "completed", "cancelled"]),
  city: zod.string().optional(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Platform summary stats
 */
export const GetStatsResponse = zod.object({
  totalArtisans: zod.number(),
  totalBookings: zod.number(),
  totalCities: zod.number(),
  totalCategories: zod.number(),
  completedJobs: zod.number(),
  averageRating: zod.number(),
});

/**
 * @summary Get top rated artisans
 */
export const getTopRatedArtisansQueryLimitDefault = 6;

export const GetTopRatedArtisansQueryParams = zod.object({
  limit: zod.coerce.number().default(getTopRatedArtisansQueryLimitDefault),
});

export const GetTopRatedArtisansResponse = zod.object({
  artisans: zod.array(
    zod.object({
      id: zod.number(),
      name: zod.string(),
      bio: zod.string().optional(),
      category: zod.string(),
      city: zod.string(),
      phone: zod.string(),
      email: zod.string().optional(),
      photoUrl: zod.string().optional(),
      priceRange: zod.string().optional(),
      yearsExperience: zod.number().optional(),
      isVerified: zod.boolean(),
      averageRating: zod.number(),
      reviewCount: zod.number(),
      createdAt: zod.coerce.date(),
    }),
  ),
  total: zod.number(),
});